Also move translations and images
