+++ shared-module/src/locales/en/course-material.json
@@ -17,6 +17,7 @@
   "close": "Close",
   "closed": "Closed!",
   "commenting-on-selection": "Commenting on: {{selection}}",
+  "congratulations": "Congratulations",
   "continue": "Continue",
   "could-not-fetch-embed": "Could not fetch {{provider}} oEmbed with URL: {{url}}",
   "course": "Course",
@@ -55,6 +56,7 @@
   "failed-to-submit": "Failed to submit {{ error }}",
   "feedback-submitted-succesfully": "Feedback submitted successfully",
   "finnish": "Finnish",
+  "generate-certicate": "Generate certificate",
   "give-feedback": "Give feedback",
   "glossary": "Glossary",
   "heres-a-list-of-all-public-pages-for-this-course": "Here's a list of all public pages for this course:",
@@ -82,6 +84,7 @@
   "proceed-to-the-next-chapter": "Proceed to the next chapter",
   "reached-end-of-course-material": "You've reached the end of the course material!",
   "reached-end-of-topic": "You've reached the end of this topic.",
+  "register": "Register",
   "screen-reader-opens-in-new-tab": "(opens in a new tab)",
   "search": "Search",
   "search-field-placeholder": "Search...",
