+++ shared-module/src/locales/fi/cms.json
@@ -1,56 +1,50 @@
 {
   "add-slide": "Lisää dia",
   "add-task": "Lisää tehtävä",
+  "block-list": "Nykyiset lohkot",
+  "block-menu": "Lisää lohko / Kaikki saatavilla olevat lohkot ",
+  "block-props": "Lohkon ominaisuudet",
   "button-text-select-media": "Valitse media (ei käytössä)",
-  "chapter-progress-placeholder": "Luvun edistyminen lohko",
   "chapter-progress-placeholder-explanation": "Tätä lohkoa käytetään luvun edistymisen näyttämiseen. Jos haluat näyttää koko kurssin edistymisen, sinun tulee käyttää Course Progress -lohkoa.",
-  "chapters-grid-placeholder": "Lukuruudukko lohko",
+  "chapter-progress-placeholder": "Luvun edistyminen lohko",
   "chapters-grid-placeholder-explanation": "Tämä lohko on sijoitettu kurssimateriaalin etusivulle, jotta voit siirtyä helposti eri lukuihin.",
+  "chapters-grid-placeholder": "Lukuruudukko lohko",
   "close": "Sulje",
-  "course-progress-placeholder": "Kurssin edistyminen lohko",
+  "congratulations-placeholder": "Onnittelulohko",
+  "congratulations-explanation": "Tämä lohko sijoitetaan kurssimateriaaliin, missä halutaan näyttää opiskelijalle onnittelut kurssin suorittamisesta. Jos kurssilta voi saada opiskelupisteitä, tämä lohko tarjoaa myös linkin sivulle, mistä opiskelija pääsee kirjaamaan suorituksensa Avoimeen yliopistoon.",
   "course-progress-placeholder-explanation": "Tämä lohko sijoitetaan kurssimateriaaliin, johon haluat näyttää kurssin edistymisen.",
+  "course-progress-placeholder": "Kurssin edistyminen lohko",
   "delete": "Poista",
   "edit": "Muokkaa",
-  "exercises-in-chapter-placeholder": "Tehtäviä luvussa lohko",
-  "exercises-in-chapter-placeholder-explanation": "Tämä lohko sijoitetaan jokaisen luvun etusivulle, esim. /luku-1/ luvun eri harjoitusten luetteloimiseen ja navigointiin.",
-  "exercise-title": "Tehtävä",
-  "exercise-name": "Tehtävän nimi",
-  "exercise-max-points": "Maksimipisteet",
-  "limit-number-of-tries": "Rajoita yrityksien lukumäärää",
-  "tries-per-slide": "Maksimimäärä yrityksiä per dia",
-  "please-select-exercise-type": "Ole hyvä ja valitse harjoitustyyppi:",
-  "error": "Virhe",
+  "editor-select-sidebar-view": "Vaihda näkymää",
   "error-cannot-render-exercise-task-missing-url": "Harjoitustehtävää ei voi renderöidä, URL-osoite puuttuu.",
   "error-spec-not-parseable": "Mallivastausta ei voida parsia",
+  "error": "Virhe",
+  "exercise-max-points": "Maksimipisteet",
   "exercise-name": "Tehtävän nimi",
   "exercise-title": "Tehtävä",
-  "exercises-in-chapter-placeholder": "Tehtäviä luvussa lohko",
   "exercises-in-chapter-placeholder-explanation": "Tämä lohko sijoitetaan jokaisen luvun etusivulle, esim. /luku-1/ luvun eri harjoitusten luetteloimiseen ja navigointiin.",
-  "glossary-placeholder": "Sanasto lohko",
+  "exercises-in-chapter-placeholder": "Tehtäviä luvussa lohko",
   "glossary-placeholder-explanation": "Tämä lohko sisältää kurssin sanaston.",
+  "glossary-placeholder": "Sanasto lohko",
   "label-email-subject": "Sähköpostin otsake",
   "label-template-name": "Sähköpostimallin nimi",
   "label-title": "Otsikko",
+  "limit-number-of-tries": "Rajoita yrityksien lukumäärää",
   "loading": "Lataa...",
-  "pages-in-chapter-placeholder": "Sivuja luvussa lohko",
   "pages-in-chapter-placeholder-explanation": "Tämä lohko sijoitetaan jokaisen luvun etusivulle, esim. /luku-1/ helpottaaksesi navigointia eri alaosioihin.",
+  "pages-in-chapter-placeholder": "Sivuja luvussa lohko",
   "please-select-exercise-type": "Ole hyvä ja valitse harjoitustyyppi:",
+  "reset": "Palauta",
   "save": "Tallenna",
   "saved": "Tallennettu",
   "saving": "Tallennetaan...",
   "serialize-to-html": "Näytä HTML-muodossa",
   "skip-to-content": "Siirry sisältöön",
   "slide-title": "Dia {{ number }}",
-  "glossary-placeholder": "Sanasto lohko",
-  "glossary-placeholder-explanation": "Tämä lohko sisältää kurssin sanaston.",
-  "skip-to-content": "Siirry sisältöön",
   "task": "Tehtävä",
   "title-assignment": "Tehtävänanto",
-  "unsupported-block-placeholder": "Ei tuettu lohko",
+  "tries-per-slide": "Maksimimäärä yrityksiä per dia",
   "unsupported-block-placeholder-explanation": "Tätä lohkoa käytetään paikkamerkkinä lohkoille, joita ei tueta.",
-  "reset": "Palauta",
-  "block-props": "Lohkon ominaisuudet",
-  "block-list": "Nykyiset lohkot",
-  "block-menu": "Lisää lohko / Kaikki saatavilla olevat lohkot ",
-  "editor-select-sidebar-view": "Vaihda näkymää"
+  "unsupported-block-placeholder": "Ei tuettu lohko"
 }
