+++ shared-module/src/locales/fi/course-material.json
@@ -17,6 +17,7 @@
   "close": "Sulje",
   "closed": "Suljettu!",
   "commenting-on-selection": "Kommentoimassa: {{selection}}",
+  "congratulations": "Onnittelut",
   "continue": "Jatka",
   "could-not-fetch-embed": "Ei voitu upottaa {{provider}} osoitteella: {{url}}",
   "course": "Course",
@@ -55,6 +56,7 @@
   "failed-to-submit": "Lähettäminen epäonnistui {{ error }}",
   "feedback-submitted-succesfully": "Palaute lisätty onnistuneesti",
   "finnish": "Suomi",
+  "generate-certicate": "Luo sertifikaatti",
   "give-feedback": "Anna palautetta",
   "glossary": "Sanasto",
   "heres-a-list-of-all-public-pages-for-this-course": "Lista kaikista julkisista sivuista tällä kurssilla",
@@ -82,6 +84,7 @@
   "proceed-to-the-next-chapter": "Siirry seuraavaan lukuun",
   "reached-end-of-course-material": "Olet päässyt kurssimateriaalin loppuun!",
   "reached-end-of-topic": "Olet saavuttanut tämän aiheen lopun",
+  "register": "Rekisteröi",
   "screen-reader-opens-in-new-tab": "(Aukeaa uudelle välilehdelle)",
   "search": "Selaa",
   "search-field-placeholder": "Selaa...",
